pyproject.toml
[tool.poetry]
name = "aiogrammer"
version = "0.1.0"
description = ""
authors = ["S1avv"]
readme = "README.md"
include = ["aiogrammer/templates/**", "aiogrammer/modules/**"]

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.19.2"}
rich = "^14.1.0"
pyyaml = "^6.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aiogrammer = "aiogrammer.cli:app"